// packages/shared/src/enums/roles.enum.d.ts
export declare enum UserRole {
    SUPER_ADMIN = "SUPER_ADMIN",
    ORG_ADMIN = "ORG_ADMIN",
    FACILITY_ADMIN = "FACILITY_ADMIN",
    DOCTOR = "DOCTOR",
    NURSE = "NURSE",
    PHARMACIST = "PHARMACIST",
    LAB_TECHNICIAN = "LAB_TECHNICIAN",
    RADIOLOGIST = "RADIOLOGIST",
    RECEPTIONIST = "RECEPTIONIST",
    BILLING_STAFF = "BILLING_STAFF",
    PATIENT = "PATIENT",
    AUDITOR = "AUDITOR"
}
export declare const UserRoleLabels: Record<UserRole, string>;
//# sourceMappingURL=roles.enum.d.ts.map